fix: link to config file examples on ampform repo
The Develop page states that AmpForm is used as an example. It's
therefore better to link to the developer config files hosted there
than the once on compwa-org (as is currently done with the :download:
role). Using :download: may be more convenient for the developer
(because one is directly notified about missing files building with
Sphinx), but for the reader the :download: is less convenient.